Move hooks into directory, update Storybook component
