+++ src/components/CustomCollection/CustomCollection.stories.tsx
@@ -3,19 +3,27 @@ import { Meta } from '@storybook/react'
 import CustomCollection from './CustomCollection'
 
 type StorybookArgTypes = {
+  handleAddBookmark: () => void
   handleRemoveBookmark: () => void
+  handleRemoveCollection: () => void
 }
 
 export default {
   title: 'Components/CustomCollection',
   component: CustomCollection,
-  argTypes: { handleRemoveBookmark: { action: 'Remove bookmark' } },
+  argTypes: {
+    handleAddBookmark: { action: 'Add bookmark' },
+    handleRemoveBookmark: { action: 'Remove bookmark' },
+    handleRemoveCollection: { action: 'Remove collection' },
+  },
 } as Meta
 
 export const ExampleCustomCollection = (argTypes: StorybookArgTypes) => (
   <CustomCollection
     title="Example collection"
     bookmarks={[{ id: 'link1', url: '#', label: 'Webmail' }]}
+    handleAddBookmark={argTypes.handleAddBookmark}
     handleRemoveBookmark={argTypes.handleRemoveBookmark}
+    handleRemoveCollection={argTypes.handleRemoveCollection}
   />
 )
